Refactored cvars into public and private sections, added mutexes, and implemented better imgui editor looks
